Integer -> Long

+++ server/src/main/java/com/server/soopool/comment/repository/BoardIdAndGroupNumberMapping.java
@@ -1,5 +1,5 @@
 package com.server.soopool.comment.repository;
 
 public interface BoardIdAndGroupNumberMapping {
-    Integer getGroupNumber();
+    Long getGroupNumber();
 }
